refactor: clean up test helpers and renamed to mock
