+++ tests/integrations/test_codex.lua
@@ -1,16 +1,13 @@
 -- Tests for OpenAI Codex integration (lua/aibo/integration/codex.lua)
 
-local helpers = require("tests.helpers")
+local mock = require("tests.mock")
 local T = require("mini.test")
 
 -- Test set
 local test_set = T.new_set({
   hooks = {
-    pre_case = function()
-      helpers.setup()
-    end,
     post_case = function()
-      helpers.cleanup()
+      vim.cmd("silent! %bwipeout!")
     end,
   },
 })
@@ -205,7 +202,7 @@ test_set["Codex check_health"] = function()
   }
 
   -- Mock executable
-  local restore = helpers.mock_executable({
+  local restore = mock.mock_executable({
     codex = true,
     node = true,
   })
